Separated logic from view.
